# docs/api/openapi/cordum-api.yaml
openapi: 3.0.3
info:
  title: Cordum HTTP API
  version: '2026-04-21.2'
  description: Canonical OpenAPI 3.0.3 spec for the Cordum gateway HTTP surface.
  contact:
    name: Cordum Team
  license:
    name: Proprietary
servers:
- url: http://localhost:8081
  description: Gateway HTTP (local default)
- url: /
  description: Relative gateway base
...
security:
- apiKey: []
- bearerAuth: []
paths:
...
  /api/v1/workflow-runs/{id}/chat:
    parameters:
    - $ref: '#/components/parameters/RunID'
    - $ref: '#/components/parameters/TenantID'
    get:
      operationId: getRunChat
      summary: Get chat messages for a run
      tags:
      - WorkflowRuns
      responses:
        '200':
          description: Chat messages
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/PaginatedResponse'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/ChatMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: postRunChat
      summary: Post a chat message to a run
      tags:
      - WorkflowRuns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostChatRequest'
      responses:
        '200':
          description: Message posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/chat:
    post:
      operationId: postLLMChat
      summary: Send a single LLM chat message
      description: Proxies an authenticated single-shot chat turn to cordum-llm-chat.
      tags:
      - LLMChat
      parameters:
      - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - message
              properties:
                session_id:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Chat turn result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierLimit'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued by the Cordum admin.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login.
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant isolation header (required on all protected routes).
      schema:
        type: string
...
    RunID:
      name: id
      in: path
      required: true
      description: Workflow run ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request — invalid or missing parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized — missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGateway:
      description: Bad gateway — upstream service error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TierLimit:
      description: Forbidden by current license entitlements
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericObject'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
      required:
      - error
      - status
    PaginatedResponse:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Cursor for the next page (null if no more pages)
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        run_id:
          type: string
        content:
          type: string
        role:
          type: string
          enum:
          - user
          - assistant
          - system
        step_id:
          type: string
          nullable: true
        job_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    PostChatRequest:
      type: object
      properties:
        content:
          type: string
        role:
          type: string
          enum:
          - user
          - assistant
          - system
        step_id:
          type: string
        job_id:
          type: string
      required:
      - content
      - role
...
    GenericObject:
      type: object
      properties: {}
      additionalProperties: true
...